# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_dormitory-life_core_internal_server_request_models.ChatMessage:
    properties:
      created_at:
        type: string
      dormitory_id:
        type: string
      email:
        type: string
      id:
        type: string
      text:
        type: string
      user_id:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.CreateChatMessageResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.CreateDormitoryResponse:
    properties:
      dormitory_id:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.DeleteDormitoryResponse:
    type: object
  github_com_dormitory-life_core_internal_server_request_models.Dormitory:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      photo_links:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.FileInfo'
        type: array
      support_email:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.FileInfo:
    properties:
      last_modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.GetChatMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ChatMessage'
        type: array
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.GetDormitoriesResponse:
    properties:
      dormitories:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.Dormitory'
        type: array
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.GetDormitoryByIdResponse:
    properties:
      dormitory:
        $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.Dormitory'
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.UpdateDormitoryResponse:
    properties:
      dormitory_id:
        type: string
    type: object
info:
  contact: {}
  description: API Core сервиса Dormitory Life
  title: Dormitory Life Core API
  version: "1.0"
paths:
  /core/dormitories:
    get:
      description: Получение списка краткой информации о всех общежитиях
      produces:
      - application/json
      responses:
        "200":
          description: Общежития
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.GetDormitoriesResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Получение списка общежитий
      tags:
      - Dormitories
    post:
      consumes:
      - application/json
      description: Создает общежитие в системе
      produces:
      - application/json
      responses:
        "200":
          description: Общежитие создано
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateDormitoryResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Создание общежития
      tags:
      - Dormitories
  /core/dormitories/{dormitory_id}:
    delete:
      description: Удаляет общежитие из системы
      produces:
      - application/json
      responses:
        "200":
          description: Общежитие удалено
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.DeleteDormitoryResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Удаление общежития
      tags:
      - Dormitories
    get:
      description: Получение подробной информации об общежитии
      produces:
      - application/json
      responses:
        "200":
          description: Общежитие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.GetDormitoryByIdResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Получение общежития
      tags:
      - Dormitories
    put:
      consumes:
      - application/json
      description: Обновляет информацию об общежитии
      produces:
      - application/json
      responses:
        "200":
          description: Общежитие обновлено
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.UpdateDormitoryResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Обновление информации об общежитии
      tags:
      - Dormitories
  /core/dormitories/{dormitory_id}/chat:
    get:
      description: Получение сообщений чата общежития
      produces:
      - application/json
      responses:
        "200":
          description: Оценки
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.GetChatMessagesResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Получение чата
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Создание сообщения в чате общежития
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение создано
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateChatMessageResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Создание cообщения
      tags:
      - Chat
...
schemes:
- http
- https
swagger: "2.0"